Implement an interface to adapt IEObjectInitialiser
for Java-related classes (all of them extend EObject) used by JaMoPP.
The main purpose of this interface is to make it easier to declare
methods for the said classes.

This interface can also be used to filter JaMoPP-related initialisers
from EObject initialisers, if initialisers for other EObject utilising
techologies are implemented in the future.