Add utility tests
These tests can be run to ensure that there are initialisers for each
Java-related element in JaMoPP.

Note: Relative paths are used within tests to access initialiser files. Therefore, if the position of cipm.consistency.initialisers.jamopp.utiltests changes relatively to the initialiser files, the said path will have to be changed for tests to work.