Implement dummy constructs for future tests
Their purpose is to be used to verify that initialisers can modify the
objects they create as intended. They can also be used to test some
static methods of IInitialiser (the top-level initialiser interface).

The said dummy constructs also include a re-usable abstract class for
further dummy constructs (to be implemented in the following commits).