Add tests for top-level similarity checking classes/interfaces
These tests make use of the dummy elements, in order to test various
interactions between top-level similarity checking mechanisms.

They also serve as a proof-of-concept for the introduced changes to
similarity checking.

The said dummy classes also contain insight about the purpose of what
they mock for tests (in form of commentary).