Add advanced tests for JaMoPP-related similarity checking
These tests aim to verify similarity checking of JaMoPP objects, whose
construction is not as simple as those from "impltests" and
"interfacetests".

The cases addressed in these tests were separated from "impltests",
because they each require a complicated setup, in order to cover certain
branches of the similarity checking logic.