Fix caseLocalVariableStatement
There are certain Commentable sub-types used by JaMoPP, which may use a Block instance, in order to store their Statement instances. If this is the case for both LocalVariableStatement (LVS) instances that are being compared, the said Block instances will be their direct container, rather than the intended one.

For example: If a LVS is to be contained within a TryBlock, it is instead contained within a Block instance first, which is then contained in the TryBlock instance (TryBlock -> Block -> LVS; not TryBlock -> LVS).

The changes made to the comparison of said containers makes sure that LVS comparison accounts for this effect. It is supposed to bypass the containers in-between, which are not the intended to be compared.