Add complex test for cyclic attributes
They dynamically explore and set each attribute A of each EObject EO of type EOT, which accepts a value of type EOT (or an array thereof), to that EObject (EO.A = EO). This makes sure that similarity checking does not run into an endless cycle and cause exceptions.

Currently disabled since similarity checking does not validate the EObjects it is comparing. Enable the test class once similarity checking can handle this.