Remove init CPRs from CPR list of PCM to Java change propagation test (experiment)
This does not change the outcome of the experiment, because those CPRs were not triggered at all

The purpose of those CPRs is to create the top-level root elements in PCM and IM models (Repository and InstrumentationModel respectively). Since the experiment copies the initial models and has to ensure that model element IDs are the same as the original tests, the init CPRs do not / should not trigger. Otherwise the model element IDs in original tests and experiment would differ, which is detrimental for Jaccard coefficient computations (for accuracy)